+++ changelog.md
@@ -1,6 +1,7 @@
 
 ## TODO
 	canvas engraver crop issue
+		turnRawSvgWithSheetDocument
 	Baking score bundle size reduction
 	implicit notes matching with sheet document
 		try articulate.ly in MIDI output
@@ -24,6 +25,8 @@
 
 
 *	2020.8.31
+
+	+	scoreMaker.ts: turnRawSvgWithSheetDocument added.
 
 	+	lilyInterpreter.ts: fixed TrackContext.processGrace duration factor.